Module sorted-collections::sortedmapExperimental

Structs

BTreeMapRange

BTreeMapRangeMut

BTreeMapRangeRemove

HashMapRange

HashMapRangeMut

HashMapRangeRemove

Traits

SortedMap
An extension trait for a Map whose keys have a defined total ordering. This trait provides convenience methods which take advantage of the map's ordering. The provided methods may be overriden if desired to provide more efficient implementations.