Trait sorted-collections::sortedmap::SortedMapExperimental

pub trait SortedMap<K, V>: Sized where K: Clone + Ord, V: Clone {
    fn first(&self) -> Option<(K, V)>;
    fn first_remove(&mut self) -> Option<(K, V)>;
    fn last(&self) -> Option<(K, V)>;
    fn last_remove(&mut self) -> Option<(K, V)>;
    fn head_map(&self, key: &K, inclusive: bool) -> Self;
    fn sub_map(&self, from_key: &K, from_inclusive: bool, to_key: &K, to_inclusive: bool) -> Self;
    fn tail_map(&self, key: &K, inclusive: bool) -> Self;

    fn ceiling(&self, key: &K) -> Option<(K, V)> { ... }
    fn floor(&self, key: &K) -> Option<(K, V)> { ... }
    fn higher(&self, key: &K) -> Option<(K, V)> { ... }
    fn lower(&self, key: &K) -> Option<(K, V)> { ... }
}

An extension trait for a Map whose keys have a defined total ordering. This trait provides convenience methods which take advantage of the map's ordering. The provided methods may be overriden if desired to provide more efficient implementations.

Required Methods

fn first(&self) -> Option<(K, V)>

Returns the first (least) key currently in this map, and its corresponding value. Returns None if this map is empty.

Examples

extern crate "sorted-collections" as sorted_collections;

use std::collections::BTreeMap;
use sorted_collections::SortedMap;

fn main() {
    let map: BTreeMap<u32, u32> =
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)].into_iter().collect();
    assert_eq!(map.first().unwrap(), (1u32, 1u32));
}

fn first_remove(&mut self) -> Option<(K, V)>

Removes and returns the first (least) key currently in this map, and its corresponding value. Returns None if this map is empty.

Examples

extern crate "sorted-collections" as sorted_collections;

use std::collections::BTreeMap;
use sorted_collections::SortedMap;

fn main() {
    let mut map: BTreeMap<u32, u32> =
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)].into_iter().collect();
    assert_eq!(map.first_remove().unwrap(), (1u32, 1u32));
    assert_eq!(map.into_iter().collect::<Vec<(u32, u32)>>(),
        vec![(2u32, 2u32), (3, 3), (4, 4), (5, 5)]);
}

fn last(&self) -> Option<(K, V)>

Returns the last (greatest) key currently in this map, and its corresponding value. Returns None if this map is empty.

Examples

extern crate "sorted-collections" as sorted_collections;

use std::collections::BTreeMap;
use sorted_collections::SortedMap;

fn main() {
    let map: BTreeMap<u32, u32> =
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)].into_iter().collect();
    assert_eq!(map.last().unwrap(), (5u32, 5u32));
}

fn last_remove(&mut self) -> Option<(K, V)>

Removes and returns the last (greatest) key currently in this map, and its corresponding value. Returns None if this map is empty.

Examples

extern crate "sorted-collections" as sorted_collections;

use std::collections::BTreeMap;
use sorted_collections::SortedMap;

fn main() {
    let mut map: BTreeMap<u32, u32> =
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)].into_iter().collect();
    assert_eq!(map.last_remove().unwrap(), (5u32, 5u32));
    assert_eq!(map.into_iter().collect::<Vec<(u32, u32)>>(),
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4)]);
}

fn head_map(&self, key: &K, inclusive: bool) -> Self

Returns the key-value pairs of this map whose keys are less than (or equal to, if inclusive is true) key, as a new instance of the same type of map.

Examples

extern crate "sorted-collections" as sorted_collections;

use std::collections::BTreeMap;
use sorted_collections::SortedMap;

fn main() {
    let map: BTreeMap<u32, u32> =
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)].into_iter().collect();

    let head_map = map.head_map(&3, true);
    // The original map shouldn't have changed.
    assert_eq!(map.into_iter().collect::<Vec<(u32, u32)>>(),
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)]);

    assert_eq!(head_map.into_iter().collect::<Vec<(u32, u32)>>(),
        vec![(1u32, 1u32), (2, 2), (3, 3)]);
}

fn sub_map(&self, from_key: &K, from_inclusive: bool, to_key: &K, to_inclusive: bool) -> Self

Returns the key-value pairs of this map whose keys range from from_key to to_key, as a new instance of the same type of map.

Examples

extern crate "sorted-collections" as sorted_collections;

use std::collections::BTreeMap;
use sorted_collections::SortedMap;

fn main() {
    let map: BTreeMap<u32, u32> =
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)].into_iter().collect();

    let sub_map = map.sub_map(&2, false, &4, true);
    // The original map shouldn't have changed.
    assert_eq!(map.into_iter().collect::<Vec<(u32, u32)>>(),
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)]);

    assert_eq!(sub_map.into_iter().collect::<Vec<(u32, u32)>>(),
        vec![(3u32, 3u32), (4, 4)]);
}

Panics

This function panics if from_key < to_key.

fn tail_map(&self, key: &K, inclusive: bool) -> Self

Returns the key-value pairs of this map whose keys are greater than (or equal to, if inclusive is true) key, as a new instance of the same type of map.

Examples

extern crate "sorted-collections" as sorted_collections;

use std::collections::BTreeMap;
use sorted_collections::SortedMap;

fn main() {
    let map: BTreeMap<u32, u32> =
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)].into_iter().collect();

    let tail_map = map.tail_map(&3, true);
    // The original map shouldn't have changed.
    assert_eq!(map.into_iter().collect::<Vec<(u32, u32)>>(),
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)]);

    assert_eq!(tail_map.into_iter().collect::<Vec<(u32, u32)>>(),
        vec![(3u32, 3u32), (4, 4), (5, 5)]);
}

Provided Methods

fn ceiling(&self, key: &K) -> Option<(K, V)>

Returns the least key greater than or equal to key, and its corresponding value. Returns None if there is no such key.

Examples

extern crate "sorted-collections" as sorted_collections;

use std::collections::BTreeMap;
use sorted_collections::SortedMap;

fn main() {
    let map: BTreeMap<u32, u32> =
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)].into_iter().collect();

    assert_eq!(map.ceiling(&3).unwrap(), (3u32, 3u32));
    // The original map shouldn't have changed.
    assert_eq!(map.into_iter().collect::<Vec<(u32, u32)>>(),
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)]);
}

fn floor(&self, key: &K) -> Option<(K, V)>

Returns the greatest key less than or equal to key, and its corresponding value. Returns None if there is no such key.

Examples

extern crate "sorted-collections" as sorted_collections;

use std::collections::BTreeMap;
use sorted_collections::SortedMap;

fn main() {
    let map: BTreeMap<u32, u32> =
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)].into_iter().collect();

    assert_eq!(map.floor(&3).unwrap(), (3u32, 3u32));
    // The original map shouldn't have changed.
    assert_eq!(map.into_iter().collect::<Vec<(u32, u32)>>(),
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)]);
}

fn higher(&self, key: &K) -> Option<(K, V)>

Returns the least key strictly greater than the given key, and its corresponding value. Returns None if there is no such key.

Examples

extern crate "sorted-collections" as sorted_collections;

use std::collections::BTreeMap;
use sorted_collections::SortedMap;

fn main() {
    let map: BTreeMap<u32, u32> =
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)].into_iter().collect();

    assert_eq!(map.higher(&3).unwrap(), (4u32, 4u32));
    // The original map shouldn't have changed.
    assert_eq!(map.into_iter().collect::<Vec<(u32, u32)>>(),
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)]);
}

fn lower(&self, key: &K) -> Option<(K, V)>

Returns the greatest key strictly less than the given key, and its corresponding value. Returns None if there is no such key.

Examples

extern crate "sorted-collections" as sorted_collections;

use std::collections::BTreeMap;
use sorted_collections::SortedMap;

fn main() {
    let map: BTreeMap<u32, u32> =
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)].into_iter().collect();

    assert_eq!(map.lower(&3).unwrap(), (2u32, 2u32));
    // The original map shouldn't have changed.
    assert_eq!(map.into_iter().collect::<Vec<(u32, u32)>>(),
        vec![(1u32, 1u32), (2, 2), (3, 3), (4, 4), (5, 5)]);
}